+++ http/test.ts
@@ -1,3 +1,4 @@
 // Copyright 2018-2019 the Deno authors. All rights reserved. MIT license.
+import "./cookie_test.ts";
 import "./server_test.ts";
 import "./file_server_test.ts";
